fixed a typo in the sse code
git-svn-id: svn://localhost:7744/trunk/alps@4965 3665984a-f8b0-4bb5-b30c-cdcaecbe7ced

+++ applications/qmc/sse4/sse_worm_prob.h
@@ -308,7 +308,7 @@ private:
 		for (unsigned en_op = 0; en_op < NOPERATORS; ++en_op) {
 			
 			double w[4] = {0.0, 0.0, 0.0, 0.0};
-			double ti_matrix[4][4];
+			unsigned ti_matrix[4][4];
 			for (unsigned i = 0; i < 4; ++i)
 				for (unsigned j = 0; j < 4; ++j)
 					ti_matrix[i][j] = EMPTY;
